Extract EduProgramViewModelBase class
