+++ R7.University/R7.University.csproj
@@ -163,6 +163,7 @@
     <Compile Include="Data\UniversityDataProvider.cs" />
     <Compile Include="ViewModels\EduProgramProfileViewModelBase.cs" />
     <Compile Include="ViewModels\EmployeeViewModelBase.cs" />
+    <Compile Include="ViewModels\EduProgramViewModelBase.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
